Support pure in-memory DB storage
Previously, we had `--storage memory` but it still required a `--db-root` directory for storing DB metadata which made it rather less useful. This PR removes the `--storage` flag and add `--db-memory` which is specified *instead* of `--db-root` and stores everything in memory.  The PR touches quite a few files but the main changes are:

  * Glean.Database.Config - support the new flag
  * Glean.Database.Storage - pull some operations into the `Storage` class
  * Glean.Catalog.Local.Memory - extract the in-memory Catalog.Store implementation from Glean.Database.Catalog.Test

Note that `--db-memory` doesn't support quite a few things (yet) but it is sufficient for quite a few things now.

Test Plan:
```
./quick.sh MODE=dev run glean:glean --db-memory --schema dir shell
[...]
> :index cpp-cmake ...
[...]
> src.File _
[...]
```