Support running multiple commands from CLI
This adds a `glean script` command which reads commands (as they would
be specified on the cmd line) from a file or stdin and executes them.
The immediate use case for me is easier benchmarking (esp. in
conjunction with `--db-memory` from #274) but this seems generally quite
useful to me and it's very easy to support.

See the example in `script --help` for how to use and how to test.